refactor: doc arena split
